+++ docs/changelog.rst
@@ -1,5 +1,10 @@
 .. currentmodule:: twitchio
 
+Master
+======
+- TwitchIO
+    - Added the ability to leave channels to the Client.
+
 2.1.4
 ======
 - TwitchIO